chore: move chat use case
