+++ lib/domain/chat/use_case/chat/get_room_detail_use_case.dart
