+++ lib/presentation/chat/chat_detail_view_model.dart
@@ -1,23 +1,23 @@
 import 'dart:async';
 import 'package:dongsoop/domain/auth/use_case/user_block_use_case.dart';
 import 'package:dongsoop/domain/chat/model/chat_message_request.dart';
-import 'package:dongsoop/domain/chat/use_case/connect_chat_room_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/disconnect_chat_room_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/get_offline_messages_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/get_paged_messages.dart';
-import 'package:dongsoop/domain/chat/use_case/get_room_detail_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/kick_user_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/leave_chat_room_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/save_chat_message_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/send_message_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/subscribe_messages_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/connect_chat_room_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/disconnect_chat_room_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/get_offline_messages_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/get_paged_messages.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/get_room_detail_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/get_user_nicknames_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/kick_user_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/leave_chat_room_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/save_chat_message_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/send_message_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/stream/subscribe_messages_use_case.dart';
 import 'package:dongsoop/presentation/chat/chat_detail_state.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:dongsoop/domain/chat/model/chat_message.dart';
 import 'package:dongsoop/providers/chat_providers.dart';
-import 'package:dongsoop/domain/chat/use_case/get_user_nicknames_use_case.dart';
 import 'package:dongsoop/core/exception/exception.dart';
-import 'package:dongsoop/domain/chat/use_case/subscribe_block_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/stream/subscribe_block_use_case.dart';
 
 class ChatDetailViewModel extends StateNotifier<ChatDetailState> {
   final ConnectChatRoomUseCase _chatRoomConnectUseCase;
