+++ lib/presentation/chat/chat_view_model.dart
@@ -1,6 +1,6 @@
 import 'package:dongsoop/core/exception/exception.dart';
 import 'package:dongsoop/domain/chat/use_case/blind_date/get_blind_date_open_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/get_chat_rooms_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/get_chat_rooms_use_case.dart';
 import 'package:dongsoop/presentation/chat/chat_state.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
