+++ lib/providers/chat_providers.dart
@@ -11,17 +11,18 @@ import 'package:dongsoop/domain/chat/use_case/blind_date/blind_disconnect_use_ca
 import 'package:dongsoop/domain/chat/use_case/blind_date/blind_send_message_use_case.dart';
 import 'package:dongsoop/domain/chat/use_case/blind_date/get_blind_session_use_case.dart';
 import 'package:dongsoop/domain/chat/use_case/blind_date/save_blind_session_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/create_QNA_chat_room_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/delete_chat_data_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/create_QNA_chat_room_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/delete_chat_data_use_case.dart';
 import 'package:dongsoop/domain/chat/use_case/blind_date/get_blind_date_open_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/get_offline_messages_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/get_paged_messages.dart';
-import 'package:dongsoop/domain/chat/use_case/get_room_detail_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/get_user_nicknames_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/get_chat_rooms_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/kick_user_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/leave_chat_room_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/save_chat_message_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/get_offline_messages_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/get_paged_messages.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/get_room_detail_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/get_chat_rooms_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/get_user_nicknames_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/kick_user_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/leave_chat_room_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/save_chat_message_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/send_message_use_case.dart';
 import 'package:dongsoop/domain/chat/use_case/stream/blind_broadcast_stream_use_case.dart';
 import 'package:dongsoop/domain/chat/use_case/stream/blind_disconnect_stream_use_case.dart';
 import 'package:dongsoop/domain/chat/use_case/stream/blind_freeze_stream_use_case.dart';
@@ -43,13 +44,12 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:dongsoop/core/network/stomp_service.dart';
 import 'package:dongsoop/core/storage/secure_storage_service.dart';
 import 'package:dongsoop/domain/chat/model/chat_message.dart';
-import 'package:dongsoop/domain/chat/use_case/connect_chat_room_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/disconnect_chat_room_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/send_message_use_case.dart';
-import 'package:dongsoop/domain/chat/use_case/subscribe_messages_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/connect_chat_room_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/chat/disconnect_chat_room_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/stream/subscribe_messages_use_case.dart';
 import 'package:dongsoop/presentation/chat/chat_detail_view_model.dart';
 import 'package:dongsoop/presentation/chat/chat_state.dart';
-import 'package:dongsoop/domain/chat/use_case/subscribe_block_use_case.dart';
+import 'package:dongsoop/domain/chat/use_case/stream/subscribe_block_use_case.dart';
 
 
 // 추후 기능, 책임 별로 providers 분리
